nyc-taxi-tips-prediction/pyproject.toml
[project]
name = "nyc-taxi-tips-prediction"
version = "0.1.0"
description = "Production-grade MLOps system to predict NYC Yellow Taxi tip amounts using the FTI pattern."
readme = "README.md"
authors = [
    { name = "Sebastian Garrido", email = "sebastiangarrido2790@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "xgboost>=2.0.0",
    "scikit-learn>=1.3.0",
    "mlflow>=2.7.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.1",
    "dvc>=3.0.0",
    "polars>=0.19.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "rich>=13.5.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.54.0",
    "plotly>=6.5.2",
    "fastapi>=0.129.0",
    "uvicorn>=0.40.0",
    "requests>=2.32.5",
    "httpx>=0.28.1",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "langchain-core>=1.2.16",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pyright>=1.1.350",
    "ipykernel>=6.25.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = ["notebooks/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "W", "B", "SIM", "C4", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.coverage.run]
omit = [
    "src/pipeline/*",
    "src/app/*",
    "main.py"
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.4",
]
